Correction: Font Path Generation.
During the initial project build, since the woff and woff2 font formats were not generated, the automatic font inclusion in stylesheets was not generated during compilation. This was due to the font conversion not being directed to the src folder when the production flag was set, resulting in no auto-generation for inclusion.